Move plugins to plugins directory
