+++ lib/plugins.cjs
@@ -6,12 +6,12 @@
 const plugin = require('./plugin.cjs');
 
 window.plugins = {
-  reference: plugin.wrap('reference', require('./reference.cjs')),
-  factory: plugin.wrap('factory', require('./factory.cjs')),
-  paragraph: plugin.wrap('paragraph', require('./paragraph.cjs')),
+  reference: plugin.wrap('reference', require('./plugins/reference.cjs')),
+  factory: plugin.wrap('factory', require('./plugins/factory.cjs')),
+  paragraph: plugin.wrap('paragraph', require('./plugins/paragraph.cjs')),
   //image: plugin.wrap('image', require './image')
-  future: plugin.wrap('future', require('./future.cjs')),
-  importer: plugin.wrap('importer', require('./importer.cjs'))
+  future: plugin.wrap('future', require('./plugins/future.cjs')),
+  importer: plugin.wrap('importer', require('./plugins/importer.cjs'))
 };
 
 // mapping between old plugins and their successor
