+++ lib/plugins/factory.cjs
@@ -2,14 +2,14 @@
 // destined to be one or another kind of item. Double click
 // will turn it into a normal paragraph.
 
-const neighborhood = require('./neighborhood.cjs');
-const plugin = require('./plugin.cjs');
-const resolve = require('./resolve.cjs');
-const pageHandler = require('./pageHandler.cjs');
-const editor = require('./editor.cjs');
-const synopsis = require('./synopsis.cjs');
-const drop = require('./drop.mjs');
-const active = require('./active.cjs');
+const neighborhood = require('../neighborhood.cjs');
+const plugin = require('../plugin.cjs');
+const resolve = require('../resolve.cjs');
+const pageHandler = require('../pageHandler.cjs');
+const editor = require('../editor.cjs');
+const synopsis = require('../synopsis.cjs');
+const drop = require('../drop.mjs');
+const active = require('../active.cjs');
 
 const escape = line => line
   .replace(/&/g, '&amp;')
