+++ lib/plugins/future.cjs
@@ -2,11 +2,11 @@
 // or wasn't found where expected. It recognizes template pages
 // and offers to clone them or make a blank page.
 
-const resolve = require('./resolve.cjs');
-const neighborhood = require('./neighborhood.cjs');
+const resolve = require('../resolve.cjs');
+const neighborhood = require('../neighborhood.cjs');
 
-const lineup = require('./lineup.cjs');
-const {rebuildPage} = require('./refresh.mjs');
+const lineup = require('../lineup.cjs');
+const {rebuildPage} = require('../refresh.mjs');
 
 const emit = function($item, item) {
