+++ lib/plugins/importer.cjs
@@ -1,8 +1,8 @@
 // An Importer plugin completes the ghost page created upon drop of a site export file.
 
-const {formatElapsedTime} = require('./util.mjs');
-const link = require('./link.mjs');
-const { newPage } = require('./page.mjs');
+const {formatElapsedTime} = require('../util.mjs');
+const link = require('../link.mjs');
+const { newPage } = require('../page.mjs');
 
 const escape = text => text
   .replace(/&/g, '&amp;')
