+++ lib/plugins/paragraph.cjs
@@ -2,9 +2,9 @@
 // with hyperlinks. It will eventually escape html tags but for the
 // moment we live dangerously.
 
-const editor = require('./editor.cjs');
-const resolve = require('./resolve.cjs');
-const itemz = require('./itemz.cjs');
+const editor = require('../editor.cjs');
+const resolve = require('../resolve.cjs');
+const itemz = require('../itemz.cjs');
 
 const type = function(text) {
   if (text.match(/<(i|b|p|a|h\d|hr|br|li|img|div|span|table|blockquote)\b.*?>/i)) {
