+++ lib/plugins/reference.cjs
@@ -3,9 +3,9 @@
 // references. Double click will edit the body of a reference
 // but not the name and site.
 
-const editor = require('./editor.cjs');
-const resolve = require('./resolve.cjs');
-const {asSlug} = require('./page.mjs');
+const editor = require('../editor.cjs');
+const resolve = require('../resolve.cjs');
+const {asSlug} = require('../page.mjs');
 
 // see http://fed.wiki.org/about-reference-plugin.html
